requirements.txt
Django==1.11.8
Fabric3==1.13.1.post1
aldryn-search==0.4.1
boto3==1.4.7
git+https://github.com/magnet-cl/django-cms.git@release/3.4.x
django-compressor-autoprefixer==0.1.0
django-compressor==2.2
django-debug-toolbar==1.8
django-filer==1.3.0
django-filter==1.1.0
django-form-utils==1.0.3
django-haystack==2.6.1
django-hitcount==1.2.3
django-libsass==0.7
django-localflavor==1.6.2
django-loginas==0.3.2
django-modeltranslation==0.12.1
django-npm==1.0.0
django-phonenumber-field==1.3.0
django-recaptcha==1.3.0
django-storages==1.6.5
djangocms-admin-style==1.2.7
djangocms-file==2.0.2
djangocms-googlemap==1.1.1
djangocms-link==2.1.2
djangocms-picture==2.0.6
djangocms-snippet==1.9.2
djangocms-style==2.0.2
djangocms-text-ckeditor==3.5.1
djangocms-video==2.0.4
djangorestframework==3.7.1
easy-thumbnails==2.4.2
git+https://github.com/Angoreher/aldryn-apphooks-config.git@master
git+https://github.com/aldryn/aldryn-newsblog.git@master
gunicorn==19.7.1
inflection==0.3.1
ipdb==0.10.3
ipython==6.0.0
markdown==2.6.9
model-mommy==1.3.2
psycopg2==2.7.1
pypugjs==4.2.2
pytz>=2017.2
requests==2.18.4
jsonpickle==0.9.5
google-api-python-client==1.6.4
beautifulsoup4==4.6.0
django-import-export==0.6.0
django-parler==1.9.1
pyopenssl==17.5.0
django-cron==0.5.0
pyasn1-modules==0.2.1
